Fixed nodemon/browsersync issues
- Using `babel-plugin-webpack-alias` with `nodemon` was causing WP to be 
loaded and restarted with the server file so I had to go nuclear and 
upgrade Babel to v7 so I could use a JS config.
  - That allowed me to dynamically build out the alias' via 
`module-resolver`. That in turn allowed me to rely souly on Babel for 
aliasing and not have to do anything in my WP or Jest configs.
- The BrowserSync issue of not always refreshing the page after a server 
restart has been resolved by building out some logic that'll ping the 
server, and call BrowserSync only after the server is up.